extract tokens in separate files
